+++ dao-jdbc-impl/src/main/java/com/itacademy/jd2/pk/hop/dao/jdbc/impl/entity/Map.java
@@ -1,5 +1,7 @@
 package com.itacademy.jd2.pk.hop.dao.jdbc.impl.entity;
 
+import java.math.BigDecimal;
+
 import com.itacademy.jd2.pk.hop.dao.api.entity.IMap;
 
 public class Map extends BaseEntity implements IMap {
@@ -7,10 +9,10 @@ public class Map extends BaseEntity implements IMap {
 	private String path;
 	private String file;
 	private Integer userId;
-	private Double latitude1;
-	private Double latitude2;
-	private Double longitude1;
-	private Double longitude2;
+	private BigDecimal latitude1;
+	private BigDecimal latitude2;
+	private BigDecimal longitude1;
+	private BigDecimal longitude2;
 
 	public String getName() {
 		return name;
@@ -44,35 +46,35 @@ public class Map extends BaseEntity implements IMap {
 		this.userId = userId;
 	}
 
-	public Double getLatitude1() {
+	public BigDecimal getLatitude1() {
 		return latitude1;
 	}
 
-	public void setLatitude1(Double latitude1) {
+	public void setLatitude1(BigDecimal latitude1) {
 		this.latitude1 = latitude1;
 	}
 
-	public Double getLatitude2() {
+	public BigDecimal getLatitude2() {
 		return latitude2;
 	}
 
-	public void setLatitude2(Double latitude2) {
+	public void setLatitude2(BigDecimal latitude2) {
 		this.latitude2 = latitude2;
 	}
 
-	public Double getLongitude1() {
+	public BigDecimal getLongitude1() {
 		return longitude1;
 	}
 
-	public void setLongitude1(Double longitude1) {
+	public void setLongitude1(BigDecimal longitude1) {
 		this.longitude1 = longitude1;
 	}
 
-	public Double getLongitude2() {
+	public BigDecimal getLongitude2() {
 		return longitude2;
 	}
 
-	public void setLongitude2(Double longitude2) {
+	public void setLongitude2(BigDecimal longitude2) {
 		this.longitude2 = longitude2;
 	}
